+++ src/app/core/mpi/components/paciente.html
@@ -102,41 +102,6 @@
                     </plex-item>
                 </plex-list>
             </div>
-            <!-- Pacientes del federador nacional-->
-            <div *ngIf='(pacientesFederador && pacientesFederador.length)'>
-                <plex-title titulo="Pacientes del Federador Nacional" size="md"></plex-title>
-                <plex-list>
-                    <plex-item *ngFor="let elem of pacientesFederador" class="">
-
-                        <plex-label *ngIf="!elem.documento" titulo="{{ elem.nombre | nombre }}"
-                                    subtitulo="{{ (elem.documento | number) || 'Sin DNI' }}">
-                        </plex-label>
-
-                        <plex-label *ngIf="elem.documento" titulo="{{ elem | nombre }}"
-                                    subtitulo="{{ elem.documento }}"></plex-label>
-
-                        <plex-label titulo="{{ elem | edad }}" subtitulo="{{ elem.fechaNacimiento | fecha:'utc' }}">
-                        </plex-label>
-
-                        <plex-label titulo="Id Federador" subtitulo="{{elem.idFederador}}">
-                        </plex-label>
-
-                        <plex-badge type="info">{{ elem | sexo }}
-                        </plex-badge>
-
-                        <plex-badge type="" class=""
-                                    [ngClass]="{'text-success' : elem.estado === 'validado' , 'text-warning' : elem.estado === 'temporal'} ">
-                            {{ elem.estado }}
-                        </plex-badge>
-
-                        <plex-badge type="success">{{ elem.percent | percent}}
-                        </plex-badge>
-
-                        <plex-button type="info" size="sm" label="Seleccionar" (click)="onSelect(elem)">
-                        </plex-button>
-                    </plex-item>
-                </plex-list>
-            </div>
             <ng-template #showPanel>
                 <paciente-panel [paciente]="pacienteModel" [direction]="detailDirection"
                                 [showRelaciones]="showRelaciones" (changeRelacion)="toRelacionesOnChange($event)">
